Refactor integration of angular velocity
This fixes a previous bug where impulses could set an object
spinning in the wrong direction.

The error came from not having settled on a convention for the
coordinate system when representing angular velocity.

The inertial tensor is calculated in body-centric coordinates
or more acurately principal coordinates. That means the inertial
tensor is diagonal.

If the rotation-vector is given in world coordinate we have to
transform the inertial tensor into world coordinates.